Update 0.0.3.0 - Added 8 Way NPC Directional Rotations
Update 0.0.3.0 adds 8 way npc directional Rotations meaning the npcs will not always be looking at the player with the looking towards player sprite, Instead it works the same as doom 1 & 2 and be able to see any direction that the npc is facing and moving towards.
Note: The walking sprite animation sprites still need to be implemented with the 8 way rotations but all the idles are done and it essentially works as intended.